% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-build.R
\name{build_panel}
\alias{build_panel}
\title{Build a country-year panel}
\usage{
build_panel(coverage, cases, demography, country_metadata = NULL)
}
\arguments{
\item{coverage}{tibble from \code{\link[=get_coverage]{get_coverage()}} (iso3, year, coverage).}

\item{cases}{tibble from \code{\link[=get_cases]{get_cases()}} (iso3, year, cases).}

\item{demography}{tibble from \code{\link[=get_demography]{get_demography()}}.}

\item{country_metadata}{Optional tibble from \code{\link[=get_country_metadata]{get_country_metadata()}} with at
least iso3, country, who_region. If provided, will be joined onto the panel.}
}
\value{
A tibble with class 'vpd_panel'.
}
\description{
Build a country-year panel
}
