% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling-odin2.R
\name{compare_methods}
\alias{compare_methods}
\title{Compare susceptibility trajectories across methods}
\usage{
compare_methods(panel, estimates, iso3)
}
\arguments{
\item{panel}{A panel with iso3/year.}

\item{estimates}{Named list of susceptibility tibbles.}

\item{iso3}{Country ISO3.}
}
\description{
Compare susceptibility trajectories across methods
}
