% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_conflict_effect_models}
\alias{fit_conflict_effect_models}
\title{Fit conflict-effect models for outbreaks and changes}
\usage{
fit_conflict_effect_models(data)
}
\arguments{
\item{data}{Output of \code{\link[=make_conflict_modelling_panel]{make_conflict_modelling_panel()}}.}
}
\value{
List containing fitted models and combined coefficient table.
}
\description{
Fit conflict-effect models for outbreaks and changes
}
