% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_outbreak_models}
\alias{fit_outbreak_models}
\title{Fit baseline outbreak models}
\usage{
fit_outbreak_models(data)
}
\arguments{
\item{data}{A modelling panel from \code{\link[=make_modelling_panel]{make_modelling_panel()}}.}
}
\value{
List with model and tidy coefficients.
}
\description{
Fit baseline outbreak models
}
