% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gho.R
\name{get_cases}
\alias{get_cases}
\title{Get reported cases series}
\usage{
get_cases(disease, indicator_code = NULL, years = NULL, countries = NULL)
}
\arguments{
\item{disease}{Disease string (e.g., "measles").}

\item{indicator_code}{Optional explicit indicator code.}

\item{years}{Optional integer vector.}

\item{countries}{Optional ISO3 vector.}
}
\value{
tibble with iso3, year, cases, cases_per_100k.
}
\description{
Get reported cases series
}
