% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gho.R
\name{gho_get}
\alias{gho_get}
\title{Low-level GET for WHO GHO OData}
\usage{
gho_get(endpoint, query = list(), cache = TRUE, quiet = TRUE)
}
\arguments{
\item{endpoint}{Character, e.g. "Indicator" or an indicator code.}

\item{query}{Named list of OData query parameters.}

\item{cache}{Logical; cache raw JSON.}

\item{quiet}{Logical.}
}
\value{
A tibble.
}
\description{
Low-level GET for WHO GHO OData
}
