% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gho.R
\name{gho_get_indicator}
\alias{gho_get_indicator}
\title{Download all data for an indicator code}
\usage{
gho_get_indicator(indicator_code)
}
\arguments{
\item{indicator_code}{Indicator code (e.g., "MCV1").}
}
\value{
tibble.
}
\description{
Download all data for an indicator code
}
