% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{make_modelling_panel}
\alias{make_modelling_panel}
\title{Create lagged modelling panel}
\usage{
make_modelling_panel(
  panel,
  suscept,
  outcome = c("outbreak_pc", "outbreak_abs"),
  lag_years = 1
)
}
\arguments{
\item{panel}{A \link{vpd_panel} including cases and population.}

\item{suscept}{Susceptibility tibble (iso3, year, susceptible_prop, susceptible_n).}

\item{outcome}{Which outbreak outcome to use.}

\item{lag_years}{Predictor lag in years.}
}
\value{
tibble.
}
\description{
Create lagged modelling panel
}
