% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{make_outcome_outbreak}
\alias{make_outcome_outbreak}
\title{Make outbreak outcome variables from cases}
\usage{
make_outcome_outbreak(cases, pop, abs_threshold = 1000, pc_threshold = 10)
}
\arguments{
\item{cases}{Numeric vector.}

\item{pop}{Numeric vector.}

\item{abs_threshold}{Absolute case count threshold.}

\item{pc_threshold}{Per-100k threshold.}
}
\value{
tibble with outbreak definitions.
}
\description{
Make outbreak outcome variables from cases
}
