% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-compat.R
\name{odin2_prepare_inputs}
\alias{odin2_prepare_inputs}
\title{Prepare odin2 inputs from panel}
\usage{
odin2_prepare_inputs(panel, times, iso3, ...)
}
\arguments{
\item{panel}{A panel with births/coverage/cases.}

\item{times}{Times (typically years).}

\item{iso3}{Country ISO3.}

\item{...}{Passed to \code{\link[=odin2_balance_inputs_from_panel]{odin2_balance_inputs_from_panel()}}.}
}
\value{
Named list of input vectors.
}
\description{
Prepare odin2 inputs from panel
}
