% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling-odin2.R
\name{odin2_simulate}
\alias{odin2_simulate}
\title{Simulate an odin2 model}
\usage{
odin2_simulate(model, times, pars = list(), initial = list(), inputs = NULL)
}
\arguments{
\item{model}{Compiled model from \code{\link[=odin2_build_model]{odin2_build_model()}}.}

\item{times}{Numeric vector.}

\item{pars}{Named list.}

\item{initial}{Named list.}

\item{inputs}{Optional named list of input vectors (for difference-equation models).}
}
\value{
tibble.
}
\description{
Simulate an odin2 model
}
