% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel-build.R
\name{panel_diagnostics}
\alias{panel_diagnostics}
\title{Extract join diagnostics from a panel}
\usage{
panel_diagnostics(panel)
}
\arguments{
\item{panel}{vpd_panel from \code{\link[=build_panel]{build_panel()}}.}
}
\value{
tibble.
}
\description{
Extract join diagnostics from a panel
}
