% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-plot.R
\name{plot_coverage_rank}
\alias{plot_coverage_rank}
\title{WHO-style coverage ranking plot}
\usage{
plot_coverage_rank(rank_tbl, top_n = 30)
}
\arguments{
\item{rank_tbl}{Output of \code{\link[=risk_rank]{risk_rank()}}.}

\item{top_n}{Number of countries to show.}
}
\description{
WHO-style coverage ranking plot
}
