% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-compat.R
\name{plot_coverage_rank_by_region}
\alias{plot_coverage_rank_by_region}
\title{Coverage ranking plot faceted by WHO region}
\usage{
plot_coverage_rank_by_region(rank_tbl, top_n = 50)
}
\arguments{
\item{rank_tbl}{Output of \code{\link[=risk_rank]{risk_rank()}}.}

\item{top_n}{Number of rows to display.}
}
\value{
A ggplot object.
}
\description{
Coverage ranking plot faceted by WHO region
}
