% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{plot_outbreak_calibration}
\alias{plot_outbreak_calibration}
\title{Calibration plot (binned)}
\usage{
plot_outbreak_calibration(predictions, n_bins = 10)
}
\arguments{
\item{predictions}{A data frame as returned by \code{evaluate_outbreak_time_split()$predictions}.}

\item{n_bins}{Number of probability bins.}
}
\value{
A ggplot.
}
\description{
Calibration plot (binned)
}
