% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{plot_outbreak_predictions_time}
\alias{plot_outbreak_predictions_time}
\title{Plot observed vs predicted outbreak probability over time}
\usage{
plot_outbreak_predictions_time(predictions)
}
\arguments{
\item{predictions}{A data frame as returned by \code{evaluate_outbreak_time_split()$predictions}.}
}
\value{
A ggplot.
}
\description{
Plot observed vs predicted outbreak probability over time
}
