% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-plot.R
\name{plot_panel_diagnostics}
\alias{plot_panel_diagnostics}
\title{Panel missingness diagnostic plot}
\usage{
plot_panel_diagnostics(panel)
}
\arguments{
\item{panel}{A \link{vpd_panel}.}
}
\description{
Panel missingness diagnostic plot
}
