% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-plot.R
\name{plot_susceptible_rank}
\alias{plot_susceptible_rank}
\title{WHO-style susceptible ranking plot}
\usage{
plot_susceptible_rank(rank_tbl, top_n = 30)
}
\arguments{
\item{rank_tbl}{Output of \code{\link[=risk_rank]{risk_rank()}}.}

\item{top_n}{Number of countries.}
}
\description{
WHO-style susceptible ranking plot
}
