% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-compat.R
\name{plot_who_style_dashboard}
\alias{plot_who_style_dashboard}
\title{WHO-style dashboard summary plot}
\usage{
plot_who_style_dashboard(rank_tbl, top_n = 30)
}
\arguments{
\item{rank_tbl}{Output of \code{\link[=risk_rank]{risk_rank()}}.}

\item{top_n}{Number of rows to display.}
}
\value{
A ggplot object.
}
\description{
WHO-style dashboard summary plot
}
