% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gho.R
\name{vpd_indicators}
\alias{vpd_indicators}
\title{Built-in mapping of common indicators}
\usage{
vpd_indicators()
}
\value{
tibble.
}
\description{
\code{vpdsus} aims to provide conservative defaults that are known to work.
Additional indicators can be discovered via \code{\link[=gho_find_indicator]{gho_find_indicator()}} and passed
explicitly to \code{\link[=get_coverage]{get_coverage()}} / \code{\link[=get_cases]{get_cases()}}.
}
