% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-package.R
\docType{package}
\name{vpdsus}
\alias{vpdsus-package}
\alias{vpdsus}
\title{vpdsus: VPD susceptibility estimation and risk scanning}
\description{
Tools to retrieve WHO GHO immunisation coverage and reported cases, join with
demography, estimate susceptibility via progressive methods (A--E), produce
WHO-style risk ranking plots, and link susceptibility to outbreaks.
}
\details{
Most users should start with:
\itemize{
\item \code{\link[=get_coverage]{get_coverage()}}, \code{\link[=get_cases]{get_cases()}}, \code{\link[=get_demography]{get_demography()}}
\item \code{\link[=build_panel]{build_panel()}}
\item \code{\link[=estimate_susceptible_static]{estimate_susceptible_static()}}, \code{\link[=estimate_susceptible_cohort]{estimate_susceptible_cohort()}}
\item \code{\link[=risk_rank]{risk_rank()}}, \code{\link[=plot_coverage_rank]{plot_coverage_rank()}}, \code{\link[=plot_susceptible_rank]{plot_susceptible_rank()}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ojwatson.github.io/vpdsus/}
  \item \url{https://github.com/OJWatson/vpdsus}
  \item Report bugs at \url{https://github.com/OJWatson/vpdsus/issues}
}

}
\author{
\strong{Maintainer}: Oliver Watson \email{ojwatson@users.noreply.github.com}

}
