% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cache.R
\name{vpdsus_cache_info}
\alias{vpdsus_cache_info}
\title{Cache information}
\usage{
vpdsus_cache_info()
}
\value{
A tibble with cache directory, number of files, and total size.
}
\description{
Cache information
}
