% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-demography.R
\name{vpdsus_example_panel}
\alias{vpdsus_example_panel}
\title{Example panel shipped with the package}
\usage{
vpdsus_example_panel()
}
\value{
A tibble.
}
\description{
Example panel shipped with the package
}
