% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gho.R
\name{gho_find_indicator}
\alias{gho_find_indicator}
\title{Search GHO indicator metadata by keyword}
\usage{
gho_find_indicator(keyword, top_n = 50)
}
\arguments{
\item{keyword}{Keyword to search within IndicatorName.}

\item{top_n}{Maximum number of results.}
}
\value{
tibble with indicator codes and names.
}
\description{
Search GHO indicator metadata by keyword
}
