% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling-odin2.R
\name{odin2_build_model}
\alias{odin2_build_model}
\title{Compile an odin2 template}
\usage{
odin2_build_model(model = c("balance", "sir", "minimal_discrete"), file = NULL)
}
\arguments{
\item{model}{One of 'balance', 'sir', or 'minimal_discrete'.}

\item{file}{Optional path to a custom template.}
}
\value{
A compiled odin2 model object.
}
\description{
Compile an odin2 template
}
